requirements.txt
asgiref==3.4.1
click==8.0.1
fastapi==0.66.0
h11==0.12.0
orjson==3.6.0
pydantic==1.8.2
python-dotenv==0.18.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
fastapi==0.68.0
h11==0.12.0
idna==3.2
pydantic==1.8.2
python-dateutil==2.8.2
python-decouple==3.4
requests==2.26.0
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
